util/io_copy: correct a file copy into an implicit relative target
The following corrects a regression with the `releng_copy` copy command
where the destination path provided is a single relative target. The
call attempts to find a "parent" directory of the target to ensure its
directory exists (to create it), but the parent directory of single
relative target is empty. The implementation failed to handle this empty
case and errors out complaining it cannot create an empty directory.
This commit corrects this logic by ignoring any directory verification
attempt if the determine parent path is empty.

Signed-off-by: James Knight <james.d.knight@live.com>