relaxed hash checking on internal development sources only
The idea of relaxed hash checking is intended for developers working
with bleeding edge implementation, where hash checking enforcement is
not desired (especially, if the sources constantly change; and there is
no desire to constantly update hashes during a development window).
Relaxed hash checking only applied if a package had configured for
development mode and releng-tool is operating in a development mode.
This meant that if a developer configured an "external" package for
development mode, there was no longer a way to perform strict hashing on
an external package if desired.

This commit tweaks the relaxed hash checking condition to only apply for
internal-flagged packages. And in the use case where a developer wishes
to ignore hashes for an external package in development mode, they may
tweak their package definition to toggle the package as an internal
package if development mode is detected.

Signed-off-by: James Knight <james.d.knight@live.com>