drop copyright year throughout the implementation
The following drops the use of including the copyright year in various
copyright statements. This prevents the need of maintaining year values
as the source evolves. Since this open source project has a public Git
repository, any individual needing to ascertain specifics of a source
file's copyright year can refer to the Git history.

The year in the LICENSE file remains as is, to help identify applicable
copyright years for distribution packages.

Signed-off-by: James Knight <james.d.knight@live.com>