tests: ensure package licenses support conjunctive/disjunctive lists
In preparation of some SDPX SBOM support, there is still a desire to
support a sequence container definition of package licenses, if a user
wishes to no place all licenses in a single string. However, using just
list types, it is hard to know if a list of licenses indicates a
conjunctive license set or a disjunctive license set. To help promote a
possible distinction -- we will treat tuple styles as conjunctive and
other types as disjunctive.

This test only aims to verify we can support processing either list and
tuple types (which is already implemented).

Signed-off-by: James Knight <james.d.knight@live.com>