introduce meson support
Provides the initial implementation to support Meson packages.

This support somewhat works. The are limitations on how releng-tool can
hint to where dependencies can be found (since if not using pkg-config
or CMake options, only compiler-related hints can be used). Also,
Windows-based install attempts appear to always fail, but this looks to
be a Meson-related issue. Ideally, we can improve support over time.

Signed-off-by: James Knight <james.d.knight@live.com>