Documentation: correct cmake define examples
When using define options in CMake, the generic examples for
`LIBFOO_*_DEFS` are not a proper representation of how CMake packages
use these defines. Instead of a direct `KEY=VAL` command line argument,
each define is prefixed with a `-D` option.

Tweaking the documentation to allow individual package types to use
either the default define example, or a define-entry example.

Signed-off-by: James Knight <james.d.knight@live.com>