cmake: use environment to hint parallel options
Instead of using CMake's `--parallel` argument, utilize the
`CMAKE_BUILD_PARALLEL_LEVEL` environment variable to hint to perform
multiple jobs for a build. This helps avoid situations where an older
version of CMake is being used and the `--parallel` argument is not
available (where using the environment is a non-issue).

With this change, we can drop the `releng.cmake.disable_parallel_option`
quirk. Also, we now sanitize the `CMAKE_BUILD_PARALLEL_LEVEL`
environment variable to prevent oddities between user-set option and
releng-tool-managed jobs.

Signed-off-by: James Knight <james.d.knight@live.com>