git: provide git repository support for build directory
The following attempts to provide Git-based sources support for
interacting with the Git repository. Typically, releng-tool keeps
extracted content separated from fetch content. If users wanting to
develop for a certain package, ideally the local-sources feature is
used. However, to be even more flexible for users, we will add a
reference to the cached repository in a `.git` file to allow performing
Git-related actions (checking for differences, working with development
branches, etc.).

This commit also adds a `releng.git.replicate_cache` quirk to perform a
full copy of the cache directory (if there are issues in some
environments pointing to a cache repository on a different partition).

Signed-off-by: James Knight <james.d.knight@live.com>