support unconfiguring development/local-sources mode
Provides options to unconfigure development mode or local-sources mode
using command line options. When the value `-` or `unset` is provided,
this will be an indication to unconfigure a given mode. This provides an
alternative way to unconfigure a mode over manually moving a file flag
or invoking a distclean.

Of course, changing a mode outside of a clean state may lead to
unexpected results.

Signed-off-by: James Knight <james.d.knight@live.com>