main: log releng-tool version by default
For every invoke of releng-tool, we will now dump the version of the
tool ahead of time. This is to help easily and quickly to inform users
of what version of the releng-tool engine they are using, which can be
helpful for users always looking to use the most recent changes.

Signed-off-by: James Knight <james.d.knight@live.com>