scripts: dump spdx license data in a formatted way
When dumping SPDX license data to an internal database, add some
identations to make the content easier to read. This is in attempt to
make it easier to reflect the changes between different license version,
at a slight cost in data size (an extra ~25KB).

This may be reverted or tweaked later.

Signed-off-by: James Knight <james.d.knight@live.com>