introduce releng script environment helpers
Adding support to provide a `RELENG_SCRIPT` and `RELENG_SCRIPT_DIR`
environment variables when processing package definitions and other
stage/custom scripts. This may make it easier for projects to quickly
reference a given script's path (or it's parent directory) without
having to rely on a series of interpreter calls.

Signed-off-by: James Knight <james.d.knight@live.com>