opts: add support for environment hint for output container
The following adds support for an `RELENG_GLOBAL_OUTPUT_CONTAINER_DIR`
environment option, which can hint to a default container folder to
store releng-tool project outputs. If set (and if an explicit output
directory is not already set), it will configure the output directory
to an folder inside the configured container folder. This should be
helpful for users building multiple releng-tool projects which has a
dedicated path/partition to manage output content.

Signed-off-by: James Knight <james.d.knight@live.com>