tox: cleanup default environments to "relaxed" ones
For tox runs, helper scripts `tox-legacy` and `tox-relaxed` provide a
way for a system to invoke various tox environments with multiple host
interpreters to test specific interpreter environments. While the
`tox-legacy` script provides a good fallback for old interpreters, there
is not specific need to needing to have a support script for newer ones.
Instead, we will just default the "working" environments in the tox
configuration (which runs the same environments originally in
`tox-relaxed`). This will not cause issues for development since a
developer's environment can run `tox-relaxed` still, and CI scripts
explicitly run specific environments.

Signed-off-by: James Knight <james.d.knight@live.com>