tool-git: prevent ssh prompts
Provide an explicit indication for SSH-based Git operations that SSH
calls should operate in a batch mode. It is desired to hard fail fast
instead of prompting a user for authentication (as many Git operations
for a single package can cause a prompt; users should have keys
provided in a user agent).

Signed-off-by: James Knight <james.d.knight@live.com>