support automatic preparation of compatible local-sources packages
When operating in a local-sources mode, releng-tool relies on internal
packages to be prepared in expected local-source container folders. If
a package cannot be found, the process stops with an error informing a
user. For new development environments, setting up local sources can be
less than ideal (especially if several internal packages are defined).
To be flexible for users, if an internal package is not found and the
tool could support setting up a local clone of a support VCS type, the
package will attempt to be initialized automatically for a user.

Signed-off-by: James Knight <james.d.knight@live.com>