only perform local-sources path check for non-default configuration
Status dumps for local-sources information will help identify if a
local-sources configuration path is not created. Although, if a
configuration uses a default state, the key's entry is not an explicit
path (`None`) which a path check would fail on. This commit updates the
check to only perform a path check if a path is set.

Signed-off-by: James Knight <james.d.knight@live.com>