adding punch action
Adding a new action "punch" which allows performing the standard
releng-tool run, but automatically trigger a re-configuration event for
any package that has already been processed.

This is similar to certain existing runs where `--force` is used, but
the force argument may change in the future. Having a specific global
action should make it a much better candidate for users who want to
operate in such a mode.

Signed-off-by: James Knight <james.d.knight@live.com>