main: default strict arguments
Updating the mainline to trigger an error event if unknown arguments are
provided. If an environment wants to have a relaxed approach for a run,
releng-tool can be now be invoked with a `--relaxed-args` argument, or
use a `RELENG_IGNORE_UNKNOWN_ARGS` environment variable to skip any
unknown argument checks.

Signed-off-by: James Knight <james.d.knight@live.com>