fetch-mercurial: handle empty cache directory
If a Mercurial fetch is aborted during its initial stages, there may be
no content generated into the cache directory. If a re-run occurs,
releng-tool will detect the (empty) cache directory exists and assume it
has already been initialized. This can cause hg to fail since it is
invoked in a way where it assumes the directory holds a valid
repository.

Adjusting the fetch process of Mercurial to also check if the cache
directory is empty. If it is, perform the same initialization steps as
if there was no directory there in the first place.

Signed-off-by: James Knight <james.d.knight@live.com>