tests: provide explicit bare-git for git testing
Tweaking Git tools testing so that for Git-related calls directly on the
Git repository, we specify a `--git-dir` to the repository. This is to
prevent issues in environments where `safe.bareRepository` is set to
`explicit`.

Another change was required to no longer use the Git repository as the
working directory as well, otherwise Git would complain about an
ambiguous argument, dealing with `HEAD` references v. the `HEAD` file in
the Git repository. Adjusting the working tree to... well, working tree
prevents this issue.

Signed-off-by: James Knight <james.d.knight@live.com>