support preallocated iterable configurations for packages
A subset of configuration entries expect a type such as a dictionary or
a list. In most common cases, users define their configuration by
creating the type and adding appropriate values:

    LIBFOO_CONF_DEFS = {'FOO': 'BAR'}

There are some cases where a user may wish to optionally add values
based on the configuration state of a package. This can require a user
to first create an initial empty type, followed by the conditional
update:

    LIBFOO_CONF_DEFS = {}

    if SomeCondition:
        LIBFOO_CONF_DEFS['FOO'] = 'BAR'

With the introduction of this commit, the package loading process will
now preallocate dictionary/list-typed configuration. This can allow a
package definition, to continue with the above example, to avoid the
need to create a configuration type before adding/appending values:

    if SomeCondition:
        LIBFOO_CONF_DEFS['FOO'] = 'BAR'

Signed-off-by: James Knight <git@jdknight.me>