packages: support packages without definition script
Packages have always explicitly required a "definition script" to be
considered a valid package. For virtual packages, this can require
having an empty definition script. To help avoid the need for an empty
definition script, check if any package scripts are defined (e.g. a
configuration or build script). If a package script is detected,
consider the folder a valid package definition.

Signed-off-by: James Knight <git@jdknight.me>