opts: utilize forwarding arguments for libfoo-exec calls
Provides an alternative fallback for arguments used in an `libfoo-exec`
call. Typically, the action will use a provided string argument to
forwarded to a shell call. Wrapping requests in a single argument can be
less than ideal when the arguments itself have quoted values. This
commit provides a fallback to use all arguments trailing `--` in the
event where no arguments are passed into default exec-pattern.

Signed-off-by: James Knight <git@jdknight.me>