flag a force environment variable when using the punch action
When users trigger a `punch` action, this should trigger rebuilds in
various packages. When `--force` was first introduced, a `RELENG_FORCE`
argument was added to help allow packages to handle scenarios when a
force-event is triggered. No flag is available for the `punch` action.
Since a `punch` aims to replace the `--force` argument, ensure this
environment flag is set to support existing configurations.

Signed-off-by: James Knight <git@jdknight.me>