util-io: ensure path-remove can remove broken symlink
The original path-remove call checks for the existence of a file before
attempting to proceed. If a symbolic link exists on a file system but
targets a non-existing path, the call immediately stops and assumes the
path has been removed. To ensure a broken symlink is removed, always
continue with the removal request in the event a broken symlink is
detected.

Signed-off-by: James Knight <git@jdknight.me>