tests: only check trailing part of path in symlink checks
A previous commit [1] attempts to avoid a symlink check issue in OSX
where a read-link would prefix with ``/private/...`. While this caused
the test to pass, it failed in Windows with a similar issue. But an
after thought, performing a realpath would break the intent of the
check when verify the contents of the symlink.

This commit reverts back to its original check, but only performing an
explicit check under Linux. For other systems, we will only enforce a
requirement that the trailing parts of the path match.

A proper fix for this can be looked at in the future.

[1]: 3eafa183f768465d6b26353bcfa8a0241f23b510

Signed-off-by: James Knight <git@jdknight.me>