util-io: ensure path-copy can copy broken symbolic links
Path copying capabilities relied on symbolic links having a valid target
to be properly copied, otherwise the copy request would fail. The
utility should support the ability to copy "broken" symbolic links that
may not have a target that exists in the build environment. Updating the
call to support this.

Signed-off-by: James Knight <git@jdknight.me>