util-io: introduce a copy-stat compatibility call
For certain file copy tests, the Python interpreter may have if a
`copystat` call is performed between two symbolic links that point to
the same file. To be flexible, check if the resulting files that are
copied are the same. If so, do not attempt to perform a copy-stat call.

Signed-off-by: James Knight <git@jdknight.me>