tests: remove edition from cargo test example
Removing the `edition` hint in the Cargo example to allow Cargo to
automatically load it into a compatible mode. The test itself does not
care so much of what edition it is versus processing the Cargo package.
In some test environments, an older version of Cargo is run and cannot
handle the original 2021 environment.

Signed-off-by: James Knight <git@jdknight.me>