scripts: improve tox-legacy execution
Adjusting the tox-legacy script to be flexible in what is executed. The
changes allow support arguments to target tox and arguments to target
releng-tool. The script also attempts to automatically detect if an
environment is set before injecting the default set, to make life easier
in adding other variables while still using the default environment set.

Signed-off-by: James Knight <git@jdknight.me>