api: introduce build base directory
Providing a hint to configuration/build/install stages to the patch of
the generic build output directory (not package specific). May be
helpful for stages which need to handle build-related options that might
need to be shared across multiple packages.

At this time, to be used in prospect Cargo work.

Signed-off-by: James Knight <git@jdknight.me>