main: only show banner on non-actioned run
While the banner is nice for logging and promotion, some action runs
that do not produce any logging makes the operation look weird (i.e.
it only outputs the version in the same manner as running
`releng-tool --version`). A user may question what has been run. To
prevent any confusion, never print the banner when an action is
provided.

Signed-off-by: James Knight <git@jdknight.me>