scripts: add coverage helper script for linux scans
Providing a helper script that can be used to invoke coverage scans in
a Linux environment using a path mounted to a Windows host. This should
allow the generation of Linux-specific coverage data that can be
populated on the Windows host, which in turn can run Windows coverage
runs and combine them into a single report.

Signed-off-by: James Knight <git@jdknight.me>