meson: only set reconfigure argument when required
This commit corrects the use of `--reconfigure` being pushed to a Meson
configuration event. First, this corrects a check on a reconfiguration
state (`RELENG_RECONFIGURE`) which would always be true (since it will
always be defined in the script environment, but set to a `True` if
enabled). Second, even if releng-tool is configured to reconfigure,
there is no need to force this flag if the output was never created for
the package.

This helps support Meson versions older than v1.1.1, which may fail when
`--reconfigure` is set without having anything to reconfigure.

Signed-off-by: James Knight <git@jdknight.me>