correct git submodule revision processing
When Git submodule logic was implemented into releng-tool, the utility
would incorrectly pull configured submodule versions. The version used
would always be the most recent revision of a submodule instead of the
fixed version defined in a supermodule.

This commit corrects the revision to pull by performing a rev-parse
check on the supermodule's submodule path.

Signed-off-by: James Knight <git@jdknight.me>