tests: correct submodule test work
This commit corrects some oddities in the initial implementation for
Git submodule (and other) testing. Submodules were missing some update
events and closing commits, as well as registering submodules early
than desired to have expected commits reference certain commits in
submodules.

In addition to the above corrections, tweaked internal commit creation
in testing to always be unique. This is to prevent the chance of
creating two commits with the same hash.

Signed-off-by: James Knight <git@jdknight.me>