support force using installer for all python setup types
Provides support for using the `installer` module when using setuptools
or distutils packages. This configures the build stage of these Python
packages to explicitly generate a wheel first (`bdist_wheel`), which can
then be processed by the `installer` module. This change allows
developers to avoid the use of `setup.py install`, since the call is
deprecated.

Signed-off-by: James Knight <git@jdknight.me>