util-io: accept a flexible execute request
Execute requests should be providing a list of arguments to be used for
a run. There can be times where a caller wishes to only invoke a single
command with no arguments, which would require a call be formatted such
as `releng_execute(['my-cmd'])`. However, adding it into a list can be
less attractive then wanting to do `releng_execute('my-cmd')`. This call
tweaks support that if a single string is provided, it can use it as
the argument.

This change also prevents providing a string that would split each
character as an argument (and eventually fail).

Signed-off-by: James Knight <git@jdknight.me>