util-io: explicit str cast on execute arguments
When processing a list of arguments for an execute call, ensure each
entry is cast to a string type before their use. This can be helpful for
environments that pass in certain arguments, such as `pathlib.Path`
values, where a caller may expect graceful support for handling the
argument type.

Signed-off-by: James Knight <git@jdknight.me>