drop enforcement of restricting a configured user-site
Builders may (or need) to rely on capabilities configured in their
user site. For example, supporting custom Python setup type modules
that can be installed in the user site but not in the system site.
This commit drops the enforcement of `PYTHONNOUSERSITE`.

Users may opt to enforce this already by configuring their project
environment options to do so.

Signed-off-by: James Knight <git@jdknight.me>