engine-fetch: allow format hints in url-mirror
Adding the ability for a package to be configured with a URL mirror that
accepts format hints. A cache site may vary on how artifacts are hosted.
They may store all content in a single directory, or content may be
restored into a container folder for an artifact type. Using the
`url_mirror` can be helpful for the former situation, but if a cache
holds package content in sub-folders, the URL mirror cannot be used as
desired. To address this, the URL mirror can be configured as a format
string, replacing the entry `name` for the name of a package (as well
as supporting `version`).

Signed-off-by: James Knight <git@jdknight.me>