support maximum version in releng-tool version check
The `releng_require_version` script helper provides a way to check a
minimum version, but no maximum version. This commit aims to support
this capability.

This change also tweaks arguments for this call to be relaxed. The
intent is to support projects defining future compatibility hints
desired for a project, but still supporting older versions of a
releng-tool engine without causing invalid argument errors.

Signed-off-by: James Knight <git@jdknight.me>