support alternative hash for a package's development revision
There may be cases where a developers defines standard hashes for a
regular release of a package that may restrict an external package from
verifying sources in a development mode build. A developer may attempt
to download latest/canary builds from upstream but still want to perform
some level of hash checking on its contents.

This commit provides a package to define an alternative hash file
definition to use for a development revision of a package. In the event
that a package has been configured for development mode a hash file
exists in the location of `libfoo.hash-<revision>`, the alternative
hash file will be used.

Signed-off-by: James Knight <git@jdknight.me>