introduce the file vcs type
Support for file URIs have been managed with the URL VCS type. The
ability to use file URIs were mainly supported with little effort since
they could be processed in the same fashion as other URL requests and
processed using `urlib`.

However, when testing in a Debian environment, it has been observed that
when if a test environment utilized `urlopen` that used a combination
of open requests both both file URIs and other URI types (e.g. http),
calls of `urlopen` would seem to always fail for non-file URIs. While
the investigation to test failures did reveal the high-level issue, the
underlying issue with the test environment was not determined. Knowing
that the use of `urlopen` calls with file URIs appeared to cause issues,
the decision was to change this handling to instead directly extract the
file paths from the site definition and perform a manual copy instead.

Signed-off-by: James Knight <git@jdknight.me>