main: promote local-sources with colon over at-symbol
The `--local-sources` supports using either `:` or `@` as a separator
for a package and its directory. The use of `:` is newer and more so
recommended since it can provide a better user experience for shell
users with completion support. Updating the documented usage in `--help`
to use the recommended separator instead.

Signed-off-by: James Knight <git@jdknight.me>