engine: allow a generic local-sources unset to unset all options
For custom modes, we want `releng-tool --development unset` and
`releng-tool --local-sources unset` to unset everything back to normal
for the respective modes. Local-sources can unset per package, but can
be a bit tedious if there are many options to set and a user does not
want to specify each path. Instead, just have the "global" site type
clear all local-sources configurations if `unset` is applied to it. This
should make a simple one-invoke call to clean up the operating mode.
This does now prevent a users from using the command line from only
unsetting the global option and leaving explicit package local-sources
options set; however, it is believed that this is a rarer use case over
trying to unset all.

Signed-off-by: James Knight <git@jdknight.me>