store fetch file-flags in their own output folder
The introduction of file flags for fetch tracking [1] had a flaw in its
approach. While it tracked the fetched state, it did store storing the
flag in a package's build output directory. This directory should not
extract before the extraction stage (with releng-tool forces it clear
before extracting, thus removing the fetch flag again). This also
generates a warning message when detected.

Instead of attempting to store this into a package's output directory,
store the file flag into a custom/new output container folder in the
build output (`.releng-tool`).

[1]: 3efeb23528a601dbfde378880700e4c87d25b9d6

Signed-off-by: James Knight <git@jdknight.me>