scripts: build releases from an lf-only checkout
Sanity check when release building that it only occurs in a setup where
the Git repository has been checked out with LF line endings.

An attempt was made to setup a Debian-based setup. The initial work
required a patch to remove the v3.13 version classifier from
`pyproject.toml`, but the patch could not be applied without complaints
about a mismatch in the patch. This is since the patch is not operating
in a binary mode and `debuild` refused to apply a fuzzy patch.

Signed-off-by: James Knight <git@jdknight.me>