util: introduce critical state utility call
Introducing a helper call `raise_for_critical` to help deal with
triggering a system exit event of calls wanting to stop the process due
to a detected critical state. This simplifies the states triggering the
system exit call.

This change also changes the original `-1` exit code to just `1`, which
may be more proper.

Signed-off-by: James Knight <git@jdknight.me>