rework alternative configuration/script priorities
The following is a rework of various "alternative files" for project
configurations and scripts. In the recent v2 release of releng-tool, the
promoted file extension is now `.rt` (e.g. `releng-tool.rt`). The long
term plan is to primarily use `.rt` and still allow a fallback to `.py`.
All other extensions, while supported for now, will now be marked as
deprecated. The priorities are as follows:

- (old) none, .rt, .releng, .py
- (new) .rt, .py, none, .releng

Signed-off-by: James Knight <git@jdknight.me>