engine-patch: always hint binary application patches
When manages patches, do not attempt to interpret line endings when
deciding on how to apply a patch. Take the patches by their raw input.

This is aimed to reduce the chance of patch failures depending on how
patches are generated, held inside a Git repository and how they are
checked out.

Signed-off-by: James Knight <git@jdknight.me>