util: replicate symlink for base links in io-copy
When a `path_copy` call replicates a directory, symbolic links inside
various folders are copied as expected. However, if a symbolic link
exists in the initial folder that is a directory, it will instead create
a new directory instead of copying the symbolic link. To correct this,
only process an entry as a directory type if its not a symbolic link.

Signed-off-by: James Knight <git@jdknight.me>