patch: from binary to verbose
A previous commit [1] attempted to always use a binary mode to be
flexible for multiple environments. However, the change did not yield
expected results as patching issues could still be observed in
environments using patch in a Strawberry Perl environment.

Oddly enough, when attempting to determine the failure for `patch`, it
had been observed that running `patch` in a `--verbose` mode alleviated
issues when applying hunks. While not ideal, but no other immediate
solution available, adjusting the original binary call/quirks to handle
a verbose flag instead.

[1]: bcd92ac2b7b4902662141d476c51dd557ac6a6b4

Signed-off-by: James Knight <git@jdknight.me>