enforce str types for sys.path changes
Force str-cast all changes to `sys.path` to be casted with `str`. This
is to aid in future work where a large portion of the implementation
will be changing from str paths to using the pathlib module. Since
`sys.path` entries are required to be strings, always enforce this even
if a provided path value is of a `Path` type.

Signed-off-by: James Knight <git@jdknight.me>