use an environment wrap for environment updates
Be more flexible when updating both the system environment and script
environments by using a wrapper that can automatically set values based
on their string representation. This is to help ready the implementation
to allow these values to be configured as Path-like objects without
needing to explicitly cast each value.

Signed-off-by: James Knight <git@jdknight.me>