relocate package configuration key-value processing
Moving the logic that verifies/finalizes package configurations out of
the large package manager file into the base package namespace. This is
to help minimize the original manager source and allow easier way to
test the validation logic.

This commit also introduces various tests to help verify this moved
implementation.

Signed-off-by: James Knight <git@jdknight.me>