do not treat the punch action as a "specific action"
When a user invokes with a `punch` action, this is more so a hint to
force trigger various stages over performing a subset of processing for
a releng-tool process. When a `punch` actions is provided, do not
internally flag the process as if there is a specific action to perform.
We want the `punch` action to do everything.

Signed-off-by: James Knight <git@jdknight.me>