use native types for environment setups
Since the environment handle now supports an environment wrapper to
automatically convert various types to strings (for `os.environ`), we no
longer need to explicitly configure various options as always string
types. Adjust booleans and integer typed options to use a better type.

Note that when detecting a `True` value for environment, we still adjust
this to `l` to keep it consistent with the original implementation (may
change over time; the existence of the flags is the check for
boolean-like environment variables).

Signed-off-by: James Knight <git@jdknight.me>