scons: run configuration stage if any configuration setting is set
The original SCons support would help trigger an optional configuration
stage if a configuration option was set. However, if a configuration
define or environment variable was set, the stage would not run. In
theory, if any configuration setting is applied to a SCons package, its
configuration stage should run; correcting.

Signed-off-by: James Knight <git@jdknight.me>