drop explicit requirement of installer module
While the installer module is required to process Python packages, it
is not required for all other use cases of releng-tool. While we still
will include the component as a dependency, attempting to be flexible
for runtime scenarios where the installer module may not be available
(but releng-tool can still run).

Signed-off-by: James Knight <git@jdknight.me>