add profile configuration support
Adding a profile configuration which provides a releng-tool explicitly
supported configuration hint dedicated for projects to use based on
their own discretion. While environment options can be passed through
or arguments forwarded after the use of `--`, it is not idiomatic as if
using arguments "directly" on the running application. This change gives
a simple means to define a "profile" that projects can use without
needing to process forward arguments. For example:

 releng-tool --profile custom-profile

What values should be used for profiles is not dictated by releng-tool
(aside from releng-tool normalizing any strings of profiles provided).
Packages are left to decide how they want to react to profile strings,
if at all. Profile arguments are no-op if packages do not attempt to
read/manage profile configurations.

Signed-off-by: James Knight <git@jdknight.me>