always output local sources paths
While a `state` request can provide details on the configured
local-sources paths, operating with a local-sources mode, it is always
nice to know what overrides paths are being used and if the configured
path exists. To keep a developer informed, always include the
configured paths when hinting at the start of a run.

Signed-off-by: James Knight <git@jdknight.me>