cmake: configure root path for find calls
For CMake configuration stages, ensure `CMAKE_FIND_ROOT_PATH` is
populated with the releng-tool-managed sysroot paths. This is to help
build against CMake toolchain definitions which may configure a
CMAKE_SYSROOT and CMAKE_FIND_ROOT_PATH default with restricted find
modes, and we want releng-tool to hint to the toolchain configuration
additional paths that should be included.

Signed-off-by: James Knight <git@jdknight.me>