pyproject.toml: drop license marking for an interim
setuptools is evolving to use the new license field definition which
accepts a SPDX-defined license identifier. Newer versions of setuptools
notes the deprecation. However, using the newer approach requires
setuptools>=75.6 to be used; and the updated approach prevents older
versions of setuptools from building the package.

For now, we will drop deprecated license approach to remove the
deprecation warning and ready ourselves for a future license marking.
When we want to decide on enforcing a minimum requirement on using
setuptools>=75.6, we can re-introduce the license hint.

Signed-off-by: James Knight <git@jdknight.me>