util-io: always process execute outputs as a stream
The use of a `poll` argument to drive either reading newlines or polling
characters was not an ideal implementation. This argument was mainly a
hack to workaround situations where command outputs that never generated
a newline (i.e. reprinting on the same line with `\r`) and the first
command output processing logic could not support this.

This commit adjusts the process to work in the output always as a stream
of data. It now reads on a larger block size but utilizing `read1` to
only block when no data is available.

A fix from this is previous calls in polling more might have had the
last line of data silently dropped if no trailing newline was provided
at the end of the last line. Any trailing data is now properly printed
and/or captured (if configured to capture).

Signed-off-by: James Knight <git@jdknight.me>