.github/check-tools.yml: handle hg path in win32
The default path for a Mercurial install with Chocolatey in placed in
`C:\Program Files\Mercurial\`. However, the path appears to not be
registered in a way that works with modern GitHub actions/latest Windows
runners. Adding the path into `GITHUB_PATH` appears to ensure `hg` is
available in future steps.

Signed-off-by: James Knight <git@jdknight.me>